Use a long for id in ssl3_get_cipher_by_id()
While the cipher id is effectively a 32-bit value, someone decided that
it should be represented by a long in various internal structs, whose
mameber is passed as id. So use a long because of this and also to make
an upcoming diff simpler.

ok jsing

+++ lib/libssl/s3_lib.c
@@ -1,4 +1,4 @@
-/* $OpenBSD: s3_lib.c,v 1.246 2023/07/08 16:40:13 beck Exp $ */
+/* $OpenBSD: s3_lib.c,v 1.247 2023/11/29 13:29:34 tb Exp $ */
 /* Copyright (C) 1995-1998 Eric Young (eay@cryptsoft.com)
  * All rights reserved.
  *
@@ -1414,7 +1414,7 @@ ssl3_get_cipher(unsigned int u)
 }
 
 const SSL_CIPHER *
-ssl3_get_cipher_by_id(unsigned int id)
+ssl3_get_cipher_by_id(unsigned long id)
 {
 	const SSL_CIPHER *cp;
 	SSL_CIPHER c;
